# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pulpoline Test API",
        "description": "Api creada para obtener los datos climaticos de ciudades",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/get-weather-by-city": {
            "post": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener clima por ciudad",
                "description": "Obtiene el clima de una ciudad determinada y registra la búsqueda. Se puede especificar el idioma de respuesta con el parámetro opcional 'lang' (valores aceptados: en, es).",
                "operationId": "1b3bf575e880094fe3d81b1d05da8f50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "description": "Nombre de la ciudad a consultar",
                                        "type": "string",
                                        "example": "Madrid"
                                    },
                                    "lang": {
                                        "description": "Código de idioma para la respuesta (opcional). Valores aceptados: en, es",
                                        "type": "string",
                                        "enum": [
                                            "en",
                                            "es"
                                        ],
                                        "example": "es"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/WeatherCityResource"
                                        },
                                        "cached": {
                                            "description": "Indica si los datos provienen de caché",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "expires_in": {
                                            "description": "Tiempo restante para que expire el caché",
                                            "type": "string",
                                            "example": "24 minutes"
                                        },
                                        "cache_status": {
                                            "description": "Estado del caché (valid, regenerated, new)",
                                            "type": "string",
                                            "example": "valid"
                                        },
                                        "is_favorite": {
                                            "description": "Indica si la ciudad está marcada como favorita",
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El parámetro city es requerido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener el clima"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/toggle-favorite": {
            "post": {
                "tags": [
                    "Weather"
                ],
                "summary": "Alternar ciudad favorita",
                "description": "Añade o elimina una ciudad de favoritos",
                "operationId": "d82b914ab299fceee8a1eb342935486e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city_id"
                                ],
                                "properties": {
                                    "city_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ciudad agregada a favoritos"
                                        },
                                        "is_favorite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciudad no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recent-searches": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener búsquedas recientes",
                "description": "Devuelve las últimas ciudades buscadas por el usuario",
                "operationId": "c0891a3659ef8ba14099c35499479c1c",
                "responses": {
                    "200": {
                        "description": "Lista de búsquedas recientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CityResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favorites": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener ciudades favoritas",
                "description": "Devuelve las ciudades marcadas como favoritas por el usuario",
                "operationId": "324cc07691f449033be8c949a311c2ff",
                "responses": {
                    "200": {
                        "description": "Lista de ciudades favoritas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CityResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CityResource": {
                "properties": {
                    "id": {
                        "description": "ID único de la ciudad",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre de la ciudad",
                        "type": "string",
                        "example": "Madrid"
                    },
                    "region": {
                        "description": "Región o estado de la ciudad",
                        "type": "string",
                        "example": "Madrid"
                    },
                    "country": {
                        "description": "País donde se encuentra la ciudad",
                        "type": "string",
                        "example": "Spain"
                    },
                    "timezone": {
                        "description": "Zona horaria de la ciudad",
                        "type": "string",
                        "example": "Europe/Madrid"
                    },
                    "created_at": {
                        "description": "Fecha de creación del registro",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "WeatherCityResource": {
                "properties": {
                    "city": {
                        "properties": {
                            "name": {
                                "description": "Nombre de la ciudad",
                                "type": "string",
                                "example": "Maturín"
                            },
                            "region": {
                                "description": "Región o estado de la ciudad",
                                "type": "string",
                                "example": "Monagas"
                            },
                            "country": {
                                "description": "País donde se encuentra la ciudad",
                                "type": "string",
                                "example": "Venezuela"
                            },
                            "tz_id": {
                                "description": "Zona horaria de la ciudad",
                                "type": "string",
                                "example": "America/Caracas"
                            }
                        },
                        "type": "object"
                    },
                    "temp_c": {
                        "description": "Temperatura en grados Celsius",
                        "type": "number",
                        "format": "float",
                        "example": 28
                    },
                    "temp_f": {
                        "description": "Temperatura en grados Fahrenheit",
                        "type": "number",
                        "format": "float",
                        "example": 82.4
                    },
                    "weather": {
                        "description": "Descripción del clima actual",
                        "type": "string",
                        "example": "Light rain shower"
                    },
                    "wind_kph": {
                        "description": "Velocidad del viento en kilómetros por hora",
                        "type": "number",
                        "format": "float",
                        "example": 15.1
                    },
                    "wind_mph": {
                        "description": "Velocidad del viento en millas por hora",
                        "type": "number",
                        "format": "float",
                        "example": 9.4
                    },
                    "humidity": {
                        "description": "Porcentaje de humedad",
                        "type": "integer",
                        "example": 100
                    },
                    "local_time": {
                        "description": "Hora local en formato epoch/unix timestamp",
                        "type": "integer",
                        "example": 1749319106
                    },
                    "parsed_localtime": {
                        "description": "Hora local formateada (DD/MM/YYYY HH:MM:SS)",
                        "type": "string",
                        "example": "07/06/2025 13:58:26"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}